feat add email notification
